+++ R/motif_enrichment.R
@@ -56,7 +56,7 @@ motif_enrichment <- function(peak_input,
     ## Handle empty input
     if (length(peak_input) == 0) {
         messager("No peaks detected in input for motif_enrichment.",
-                "Returning empty result.", v = verbose, parallel = TRUE)
+                "Returning empty result.", v = verbose)
         return(empty_result)
     }
     
@@ -74,7 +74,7 @@ motif_enrichment <- function(peak_input,
     ## Handle zero enrichment
     if (is.null(ame_out)) {
         messager("No peaks were enriched for the input motif.",
-                "Returning empty result.", v = verbose, parallel = TRUE)
+                "Returning empty result.", v = verbose)
         return(empty_result)
     }